Move User type to data package
