+++ packages/services/src/permissions/useUserCapabilities.ts
@@ -1,8 +1,9 @@
 import { useMemoStringify } from '@eventespresso/hooks';
+import type { User } from '@eventespresso/data';
 
-import { useConfig, CurrentUserProps } from '../config';
+import { useConfig } from '../config';
 
-const useUserCapabilities = (): CurrentUserProps['capabilities'] => {
+const useUserCapabilities = (): User['capabilities'] => {
 	const { currentUser } = useConfig();
 
 	return useMemoStringify(currentUser?.capabilities || []);
